// frontend/src/types/User.ts
export type User = {
  createdAt: string;
  email: string;
  isActive: boolean;
  name: string;
  role: string;
  updatedAt: string;
  _id: string;
  profilePic: string;
  password: string;
  __v: number;
};

export type allUsers = {
  address: string;
  age: string;
  createdAt: string;
  gender: string;
  name: string;
  skills: string;
  updatedAt: string;
  _id: string;
  __v: number;
};